Fixed warning about size of pointer.

+++ desktop/src/mesh/omMesh.cpp
@@ -115,7 +115,7 @@ void OmMesh::makeDisplayList(QGLContext const* context)
     glBindBufferARB(GL_ARRAY_BUFFER_ARB, vertexDataVboId_);
 
     //specify vector size for interleaved vector data
-    uint32_t vector_size = 3 * sizeof(GL_FLOAT);
+    size_t vector_size = 3 * sizeof(GL_FLOAT);
 
     //specify normal (type, stride, pointer)
     glNormalPointer(GL_FLOAT, 2 * vector_size, (void *)vector_size);
